+++ src/Resources/contao/dca/tl_ls_shop_variant.php
@@ -1402,8 +1402,24 @@ class tl_ls_shop_variant_controller extends Backend {
 	}
 
 	public function insertAttributeValueAllocationsInAllocationTable($str_value, DataContainer $dc) {
-		ls_shop_generalHelper::insertAttributeValueAllocationsInAllocationTable(json_decode($str_value), $dc->id, 1);
-		return $str_value;
+
+        $data = json_decode($str_value, true);
+
+        //if json_decode returns no valid array, this normally happens if str_value ist empty
+        if(!is_array($data))
+        {
+            $data = array();
+        }
+
+        //If an attribute does not exist, remove it so that it is not saved.
+        $data = array_filter($data, function($item) {
+            return !empty($item[1]);
+        });
+
+		ls_shop_generalHelper::insertAttributeValueAllocationsInAllocationTable($data, $dc->id, 1);
+
+        //return array with removed empty attributes
+        return json_encode(array_values($data));
 	}
 
 	public function listVariants($arrRow) {