+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Leading Systems Contao Merconis bundle changelog
  * add hook 'validateCoupon'
  * improve product images identification
  * fix customizer instanz
+ * fix darkmode backend
 
 ### 5.1.2 (2025-04-11)
  * add indices