+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Leading Systems Contao Merconis bundle changelog
 ===========================================
 
+### 5.1.3 (2025-05-09)
+ * add hook 'validateCoupon'
+ * improve product images identification
+ * fix customizer instanz
+ * fix darkmode backend
+ * fix insertAttributeValueAllocationsInAllocationTable()
+
 ### 5.1.2 (2025-04-11)
  * add indices
  * fix LiveHits